Ajout du pathfinding A* sur les mouvements de suivi des ennemis vers le joueur

+++ Project1/Node.hpp
@@ -4,9 +4,9 @@
 
 struct Node {
     sf::Vector2i position;
-    int gCost, hCost, fCost;
+    float gCost, hCost, fCost;
     Node* parent;
 
     Node(sf::Vector2i pos);
-    void calculateCosts(Node* end, int newG);
+    void calculateCosts(Node* end, float newG);
 };
